models/requirements.txt
psycopg2-binary==2.9.4
sqlalchemy==1.4.41
pandas==1.5.0
scikit-learn==1.1.2
flake8==5.0.4
black==22.10.0
scikit-surprise==1.1.3
cachetools==5.2.0
wheel==0.37.1
pyspark==3.3.0
python-dotenv==0.21.0
prometheus-client==0.15.0
requests==2.28.1
gunicorn[gevent]
tensorboard==2.11.2
matplotlib==3.6.3
pika==1.3.1
redis==4.5.1
socket.io-emitter==0.1.5.1
aioredis[hiredis]==2.0.1
aiohttp==3.8.4
